old adafactor setup for hnet

+++ hyper_task_descriptions/modeling/hyper_transformer.py
@@ -494,7 +494,7 @@ class HyperEncoderDecoderModel(EncoderDecoderModel):
                 [instruction_embedding, encoded], axis=1
             )
             inputs = jnp.concatenate(
-                [hyper_inputs, dtype=jnp.int32), inputs], axis=1
+                [hyper_inputs, inputs], axis=1
             )
         encoded_inputs = decoding.flat_batch_beam_expand(
             encoded,
